Query search results using # fragment
