+++ blocks/header/header.js
@@ -1,14 +1,7 @@
 import handleViewportChanges from './header-events.js';
 import { getMetadata, decorateIcons, toClassName } from '../../scripts/lib-franklin.js';
 import { buildHamburger } from './menus/mobile-menu.js';
-import {
-  div,
-  li,
-  h3,
-  form,
-  input,
-  button,
-} from '../../scripts/dom-helpers.js';
+import buildSearch from './menus/search.js';
 
 function buildBrandLogo(content) {
   const logoWrapper = document.createElement('div');
@@ -31,48 +24,6 @@ function buildRequestQuote() {
   requestQuote.classList.add('header-rfq');
   requestQuote.innerHTML = '<a title="" href="/quote-request?cid=12" target="">Request<br>Quote</a>';
   return requestQuote;
-}
-
-function buildSearch(content) {
-  // get div with class Robot Image with Speach from navContent
-  const robotDiv = content.querySelector('.robot-image-with-speech');
-
-  const search = li(
-    { class: 'searchlink header-search fa fa-search', 'aria-expanded': 'false' },
-    div(
-      { class: 'menu-nav-search-flex' },
-      div(
-        { class: 'menu-nav-search-view' },
-        robotDiv,
-        div(
-          { class: 'menu-nav-search-bar' },
-          h3(
-            'Search',
-          ),
-          div(
-            { class: 'search-form-group' },
-            form(
-              { action: '/search-results', method: 'GET' },
-              input(
-                {
-                  id: 'search_keyword_search1', class: 'form-control', type: 'text', placeholder: 'moleculardevices.com', name: 'search',
-                },
-              ),
-              button(
-                { class: 'transparentBtn btn searchbutton', type: 'submit' },
-                'Search',
-              ),
-            ),
-          ),
-        ),
-      ),
-      div(
-        { class: 'menu-nav-submenu-close' },
-      ),
-    ),
-  );
-
-  return search;
 }
 
 export async function fetchHeaderContent() {
